backend/requirements.txt
Flask
Flask-CORS
python-dotenv
deepgram-sdk
